further improves the accuracy of page size calculation

+++ src/main/java/uk/co/mruoc/html/HtmlParser.java
@@ -12,7 +12,7 @@ public class HtmlParser {
     private static final String TITLE_SELECTOR = ".productTitleDescriptionContainer h1";
     private static final String PRICE_SELECTOR = "p.pricePerUnit";
     private static final String DESCRIPTION_SELECTOR = "div.productText";
-    private static final int BYTES_IN_KILOBYTE = 1024;
+    private static final double BYTES_IN_KILOBYTE = 1024;
 
     private final String html;
     private final Document document;
